requirements.txt
flask==3.0.*
flask-wtf==1.2.*
flask-sqlalchemy==3.0.*
flask-migrate==3.0.*
flask-login==0.6.3
email-validator==2.3.0
flask-mail==0.9.1
pyjwt==2.8.0
python-dotenv==1.0.1
flask-moment==1.0.5
flask-babel
langdetect
requests
google-genai